Switch to hatchling build
